Working on enemy reactive 3D audio system.
It's only bare-bones right now. Soon I'll adjust the enemy AI to react
to the volume of the sounds, thought it'll be difficult in that it might
need an audio listener for volume detection, but not to actually send
anything to the speakers, and unity doesn't like more than 1 audio
listener at a time.