A bunch of fixes, more to come...
- moved trap prefabs into trap folder

- lowered jump height of player and fixed another jump bug. The bug
allowed the player to jump over walls. This was caused from the walls
not having rigidbodies. It turns out that for collisions between the
player and objects to work properly, those objects need rigidbodies...
so I attached rigidbodies to everything, however they are kinetic so
that they won't move due to the player colliding with them or gravity.

- player still needs a little more work with the camera (doing that
next)

- increased hitbox on torches so the player can turn them on easier