Separating Axis Theorem used
SAT is working perfectly to determine the instant that the player's
collision box touches the end tile. Right now it's simply creating a
debug statement to indicate that the end has been reached.

To see the corners and normals of both the player and the end tile,
there are several Debug.DrawRay statements in DetectPlayer_SAT.cs that
can simply be uncommented.