Now we have random traps.
In MazeGenerator.cs there are now 2 public lists. One for wallTraps and
one for floorTraps. If we ever create more traps, just drag your trap
gameobject onto the public list in unity.

Now I just have to adjust the rotation of Ryan's trap and we should be
good for tomorrow.