Had to add a rigidbody to pendulum..
my bad, forgot to tell you guys that the object that has the TrapDamage
script must have a box collider as well as a rigidbody. If you don't
want the rigidbody to do anything, make it Is Kinematic as well as
freeze all positions and rotations and disable gravity.

I also added this to Ryan's arrow/ball in a previous commit.